requirements.txt
IMDbPY==2022.7.9
pymediainfo==5.1.1.dev1+g2e5891b
